correction erreurs follow line

+++ assess_environment.h
@@ -1,9 +1,11 @@
 #ifndef ASSESS_ENVIRONMENT_H
 #define ASSESS_ENVIRONMENT_H
 
-
-int get_line_position(void);
+uint16_t get_line_position(void);
 int get_line_number(void);
 void assess_environment_start(void);
+//int process_line_number(void);
+//uint16_t extract_line_width(uint8_t *buffer);
+
 
 #endif /* ASSESS_ENVIRONMENT_H */
